# dev/transit-back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation de l'API Transit",
        "description": "Documentation de l'API trans.it générée avec Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/documents/{id}/link": {
            "get": {
                "tags": [
                    "Liens"
                ],
                "summary": "Récupérer le lien associé à un document",
                "operationId": "051544562b2ccfc265baa48587efa669",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du document",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lien trouvé"
                    },
                    "404": {
                        "description": "Aucun lien trouvé pour ce document"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/links": {
            "get": {
                "tags": [
                    "Liens"
                ],
                "summary": "Récupérer tous les liens",
                "operationId": "faaabb479740384b02ad5fcdaa99a163",
                "responses": {
                    "200": {
                        "description": "Liste de tous les liens"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Liens"
                ],
                "summary": "Créer un lien de partage",
                "operationId": "a739b6c1e5b762fee5fc0a148f4e2682",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Id_Document",
                                    "Code_unique",
                                    "URL"
                                ],
                                "properties": {
                                    "Id_Document": {
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "Code_unique": {
                                        "type": "string",
                                        "example": "ABC123XYZ"
                                    },
                                    "URL": {
                                        "type": "string",
                                        "example": "https://trans.it/download/ABC123XYZ"
                                    },
                                    "Nb_Acces": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "IsActive": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "Date_Expiration": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-12-31 23:59:59"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lien créé avec succès"
                    },
                    "400": {
                        "description": "Erreur lors de la création"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/links/{id}": {
            "put": {
                "tags": [
                    "Liens"
                ],
                "summary": "Mettre à jour un lien de partage",
                "operationId": "5a5848eb3f046cc360a65da7738c813f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du lien à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Code_unique": {
                                        "type": "string",
                                        "example": "NEWCODE999"
                                    },
                                    "URL": {
                                        "type": "string",
                                        "example": "https://trans.it/download/NEWCODE999"
                                    },
                                    "Nb_Acces": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "IsActive": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "Date_Expiration": {
                                        "type": "string",
                                        "example": "2026-01-01 00:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lien mis à jour"
                    },
                    "400": {
                        "description": "Erreur lors de la mise à jour"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Liens"
                ],
                "summary": "Supprimer un lien de partage",
                "operationId": "51d72b7644263612fc13cbd1f01a7b0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du lien",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lien supprimé"
                    },
                    "400": {
                        "description": "Erreur lors de la suppression"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}/links": {
            "get": {
                "tags": [
                    "Liens"
                ],
                "summary": "Récupérer tous les liens d’un utilisateur",
                "operationId": "09f8b83a1fb3d7f757082309eadac810",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des liens associés à l'utilisateur"
                    },
                    "404": {
                        "description": "Aucun lien trouvé ou utilisateur inexistant"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}